changes for requested fixes
