+++ tensorboard/plugins/text/text_demo.py
@@ -1,4 +1,4 @@
-# Copyright 2017 The TensorFlow Authors. All Rights Reserved.
+# Copyright 2020 The TensorFlow Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,20 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Sample text summaries exhibiting all the text plugin features."""
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 from absl import app
 from absl import logging
-from six.moves import xrange  # pylint: disable=redefined-builtin
 import tensorflow as tf
+from tensorboard.plugins.text.summary_v2 import text
 
-from tensorboard.util import tb_logging
-
-logger = tb_logging.get_logger()
+tf.compat.v1.enable_v2_behavior()
 
 # Directory into which to write tensorboard data.
 LOGDIR = "/tmp/text_demo"
@@ -39,8 +32,24 @@ def simple_example(step):
     # UTF-8. Here's a simple example, wherein we greet the user on each
     # step:
     step_string = tf.as_string(step)
-    greeting = tf.strings.join(["Hello from step ", step_string, "!"])
-    tf.compat.v1.summary.text("greeting", greeting)
+    greeting = tf.constant("hello from step %d! 😊" % step)
+    text("greeting", greeting, step)
+
+
+def simple_example_with_pagination(step):
+    # Text summaries can be grouped by Tag (the first argument to the `text`
+    # summary's API).  Unique tags will have their own dropdown in the frontend,
+    # unless those tags are sub-tags of another tag.  To create a subtag name,
+    # one would input MainTagName/SubTagName as the input argument to the `text`
+    # summary API.  Subtags will be displayed within the main tag's dropdown,
+    # and a pager will appear if there are more than 12 subtags.  This test
+    # displays that UI feature by creating 36 subtags within the main tag:
+    # "simple_example_with_pagination"
+    for i in range(36):
+        step_string = tf.as_string(step)
+        greeting = tf.constant("hello from step %d! 😊" % step)
+        tag = "card%d/greeting" % (i + 1)
+        text(tag, greeting, step)
 
 
 def markdown_table(step):
@@ -65,7 +74,7 @@ def markdown_table(step):
     table = tf.strings.join([header_row, "---|---", table_body], separator="\n")
     preamble = "We conducted an experiment and found the following data:\n\n"
     result = tf.strings.join([preamble, table])
-    tf.compat.v1.summary.text("chocolate_study", result)
+    text("chocolate_study", result, step)
 
 
 def higher_order_tensors(step):
@@ -100,35 +109,38 @@ def higher_order_tensors(step):
     # The result, `table_full`, is a rank-2 string tensor of shape
     # `[step + 1, step + 1]`. We can pass it directly to the summary, and
     # we'll get a nicely formatted table in TensorBoard.
-    tf.compat.v1.summary.text("multiplication_table", table_full)
+    text("multiplication_table", table_full, step)
+
+
+def create_run(logdir, run_name):
+    # Creates a run specified by `run_name` populated with the text summaries
+    # generated by the function above.
+    writer = tf.summary.create_file_writer(logdir + "/" + run_name)
+
+    with writer.as_default():
+        for step in range(STEPS):
+            with tf.name_scope("simple_example"):
+                simple_example(step)
+            with tf.name_scope("markdown_table"):
+                markdown_table(step)
+            with tf.name_scope("higher_order_tensors"):
+                higher_order_tensors(step)
+            with tf.name_scope("simple_example_with_pagination"):
+                simple_example_with_pagination(step)
+
+    writer.close()
 
 
 def run_all(logdir):
-    tf.compat.v1.reset_default_graph()
-    step_placeholder = tf.compat.v1.placeholder(tf.int32)
-
-    with tf.name_scope("simple_example"):
-        simple_example(step_placeholder)
-    with tf.name_scope("markdown_table"):
-        markdown_table(step_placeholder)
-    with tf.name_scope("higher_order_tensors"):
-        higher_order_tensors(step_placeholder)
-    all_summaries = tf.compat.v1.summary.merge_all()
-
-    with tf.compat.v1.Session() as sess:
-        writer = tf.summary.FileWriter(logdir)
-        writer.add_graph(sess.graph)
-        for step in xrange(STEPS):
-            s = sess.run(all_summaries, feed_dict={step_placeholder: step})
-            writer.add_summary(s, global_step=step)
-        writer.close()
+    create_run(logdir, "run1")
+    create_run(logdir, "run2")
 
 
 def main(unused_argv):
     logging.set_verbosity(logging.INFO)
-    logger.info("Saving output to %s." % LOGDIR)
+    logging.info("Saving output to %s." % LOGDIR)
     run_all(LOGDIR)
-    logger.info("Done. Output saved to %s." % LOGDIR)
+    logging.info("Done. Output saved to %s." % LOGDIR)
 
 
 if __name__ == "__main__":
